Rename base files
Just gets rid of the redundant base-actor, base-item etc.
