+++ src/module/actor/index.js
@@ -1,4 +1,4 @@
-export { TorchbearerBaseActor } from "./base-actor";
+export { TorchbearerBaseActor } from "./base";
 export { TorchbearerCharacterActor, TorchbearerCharacterSheet } from "./character";
 export { TorchbearerMonsterActor, TorchbearerMonsterSheet } from "./monster";
 export { TorchbearerNPCActor, TorchbearerNPCSheet } from "./npc";
