+++ src/module/item/index.js
@@ -1,3 +1,3 @@
-export { TorchbearerBaseItem } from "./base-item";
+export { TorchbearerBaseItem } from "./base";
 export { TorchbearerItem, TorchbearerItemSheet } from "./item";
 export { TorchbearerSpell, TorchbearerSpellSheet } from "./spell";
