+++ src/module/item/item/sheet.js
@@ -1,4 +1,4 @@
-import { TorchbearerBaseItemSheet } from "../base-item-sheet";
+import { TorchbearerBaseItemSheet } from "../base-sheet";
 
 export class TorchbearerItemSheet extends TorchbearerBaseItemSheet {
   /** @override */
