+++ src/module/item/spell/item.js
@@ -1,3 +1,3 @@
-import { TorchbearerBaseItem } from "../base-item";
+import { TorchbearerBaseItem } from "../base";
 
 export class TorchbearerSpell extends TorchbearerBaseItem {}
